requirements.txt
flask==2.2.3
requests==2.28.2
flask_cors==3.0.10
python-dotenv==0.21.1